Revert "移动功能，正则修改，重命名修改"
This reverts commit f66a7e903369b6e2adf04574e2fc52a4e871866d.

+++ sastdisk.sql
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS `disk_file` (
   `time` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `hash` char(40) NOT NULL,
   `is_dir` tinyint(1) NOT NULL,
-  `deleted` int(11) NOT NULL DEFAULT '0',
+  `deleted` tinyint(1) NOT NULL DEFAULT '0',
   `filesize` int(11) NOT NULL,
   PRIMARY KEY (`fid`),
   KEY `uid` (`uid`,`deleted`)
